arreglo que width y height no sean un x2 e y2

+++ server_src/headers/personaje.h
@@ -15,7 +15,7 @@ class ListaObjetos;
 
 class Bala: public Objeto {
 protected:
-    int vel;
+    float vel;
 
 public:
     int danio;
